feat: Create the functionalities to manage the http response and request
BREAKING CHANGE: configure missing http methods.
